requirements.txt
Django==1.6.5
South==1.0
argparse==1.2.1
defusedxml==0.4.1
django-manager-utils==0.6.0
django-query-builder==0.5.6
django-tastypie==0.10.0
fleming==0.4
lxml==3.3.5
mimeparse==0.1.3
psycopg2==2.5.3
python-dateutil==2.2
pytz==2014.4
six==1.7.3
tweepy==2.3.0
wsgiref==0.1.2
